requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
paho-mqtt==1.4.0
python-logstash==0.4.6
raven==6.9.0
